Commit Feature Chat Bot
- Membuat halaman
- Promping sederhana
- Menambahkan halaman di route
